Crate rattler_shell

This crate provides helper functions to activate and deactivate virtual environments.

Re-exports

pub use run::run_in_environment;

Modules

activation

This crate provides helper functions to activate and deactivate virtual environments.

run

Helpers to run commands in an activated environment.

shell

This module contains the Shell trait and implementations for various shells.