rattler_shell

Module activation

This crate provides helper functions to activate and deactivate virtual environments.

Structs

ActivationResult

The result of a activation. It contains the activation script and the new path entries. The activation script already sets the PATH environment variable, but for “environment stacking” purposes it’s useful to have the new path entries separately.

ActivationVariables

A struct that contains the values of the environment variables that are relevant for the activation process. The values are stored as strings. Currently, only the PATH and CONDA_PREFIX environment variables are used.

Activator

A struct that holds values for the activation and deactivation process of an environment, e.g. activation scripts to execute or environment variables to set.

Enums

ActivationError

Error that can occur when activating a conda environment

PathModificationBehavior

Type of modification done to the PATH variable

Functions

prefix_path_entries

Return a vector of path entries that are prefixed with the given path.