rattler_shell::activation

Struct Activator

pub struct Activator<T: Shell + 'static> {
    pub target_prefix: PathBuf,
    pub shell_type: T,
    pub paths: Vec<PathBuf>,
    pub activation_scripts: Vec<PathBuf>,
    pub deactivation_scripts: Vec<PathBuf>,
    pub env_vars: IndexMap<String, String>,
    pub post_activation_env_vars: IndexMap<String, String>,
    pub platform: Platform,
}

A struct that holds values for the activation and deactivation process of an environment, e.g. activation scripts to execute or environment variables to set.

Fields
target_prefix: PathBuf

The path to the root of the conda environment

shell_type: T

The type of shell that is being activated

paths: Vec<PathBuf>

Paths that need to be added to the PATH environment variable

activation_scripts: Vec<PathBuf>

A list of scripts to run when activating the environment

deactivation_scripts: Vec<PathBuf>

A list of scripts to run when deactivating the environment

env_vars: IndexMap<String, String>

A list of environment variables to set before running the activation scripts. These are evaluated before activation_scripts have run.

post_activation_env_vars: IndexMap<String, String>

A list of environment variables to set after running the activation scripts. These are evaluated after activation_scripts have run.

platform: Platform

The platform for which to generate the Activator

Implementations

impl<T: Shell + Clone> Activator<T>

pub fn from_path( path: &Path, shell_type: T, platform: Platform, ) -> Result<Activator<T>, ActivationError>

Create a new activator for the given conda environment.

Arguments

path - The path to the root of the conda environment

shell_type - The shell type that the activator is for

operating_system - The operating system that the activator is for

Returns

A new activator

Examples

use rattler_shell::activation::Activator;
use rattler_shell::shell;
use rattler_conda_types::Platform;
use std::path::PathBuf;

let activator = Activator::from_path(&PathBuf::from("tests/fixtures/env_vars"), shell::Bash, Platform::Osx64).unwrap();
assert_eq!(activator.paths.len(), 1);
assert_eq!(activator.paths[0], PathBuf::from("tests/fixtures/env_vars/bin"));

pub fn activation( &self, variables: ActivationVariables, ) -> Result<ActivationResult<T>, ActivationError>

Create an activation script for a given shell and platform. This returns a tuple of the newly computed PATH variable and the activation script.

pub fn deactivation( &self, variables: ActivationVariables, ) -> Result<ActivationResult<T>, ActivationError>

Create a deactivation script for the environment. This returns the deactivation script that unsets environment variables and runs deactivation scripts.

pub fn run_activation( &self, variables: ActivationVariables, environment: Option<HashMap<&OsStr, &OsStr>>, ) -> Result<HashMap<String, String>, ActivationError>

Runs the activation script and returns the environment variables changed in the environment after running the script.

If the environment parameter is not None, then it will overwrite the parent environment variables when running the activation script.