rattler_shell

Module shell

This module contains the Shell trait and implementations for various shells.

Structs

Bash

A Shell implementation for the Bash shell.

CmdExe

A Shell implementation for the cmd.exe shell.

Fish

A Shell implementation for the Fish shell.

NuShell

A Shell implementation for the Bash shell.

ParseShellEnumError

Parsing of a shell was not possible. The shell most likely is not supported.

PowerShell

A Shell implementation for PowerShell.

ShellScript

A helper struct for generating shell scripts.

Xonsh

A Shell implementation for the Xonsh shell.

Zsh

A Shell implementation for the Zsh shell.

Enums

ShellEnum

A generic Shell implementation for concrete shell types.

ShellError

An error that can occur when working with shell scripts.

Traits

Shell

A trait for generating shell scripts. The trait is implemented for each shell individually.