rattler_shell::shell

Struct ShellScript

pub struct ShellScript<T: Shell> { /* private fields */ }

A helper struct for generating shell scripts.

Implementations

impl<T: Shell + 'static> ShellScript<T>

pub fn new(shell: T, platform: Platform) -> Self

Create a new ShellScript for the given shell.

pub fn apply_env_vars_with_backup( &mut self, current_env: &HashMap<String, String>, new_shlvl: i32, envs: &IndexMap<String, String>, ) -> Result<&mut Self, ShellError>

Apply the provided environment variables to the script while backing up existing values to the current shell level.

pub fn set_env_var( &mut self, env_var: &str, value: &str, ) -> Result<&mut Self, ShellError>

Export an environment variable.

pub fn unset_env_var(&mut self, env_var: &str) -> Result<&mut Self, ShellError>

Unset an environment variable.

pub fn set_path( &mut self, paths: &[PathBuf], path_modification_behavior: PathModificationBehavior, ) -> Result<&mut Self, ShellError>

Set the PATH environment variable to the given paths.

pub fn run_script(&mut self, path: &Path) -> Result<&mut Self, ShellError>

Run a script in the generated shell script.

pub fn source_completions( &mut self, completions_dir: &Path, ) -> Result<&mut Self, ShellError>

Source completion scripts for the shell from a given directory with completion scripts.

pub fn append_script(&mut self, script: &Self) -> &mut Self

Add contents to the script. The contents will be added as is, so make sure to format it correctly for the shell.

pub fn contents(&self) -> Result<String, ShellError>

Return the contents of the script.

pub fn print_env(&mut self) -> Result<&mut Self, Error>

Print all environment variables to stdout during execution.

pub fn echo(&mut self, text: &str) -> Result<&mut Self, Error>

Run echo in the shell script.

pub fn restore_env_var( &mut self, key: &str, backup_key: &str, ) -> Result<&mut Self, ShellError>

Restores an environment variable from its backup if it exists, otherwise unsets it.